fixed method {add to redirect function in restrict_access_json}

+++ app/controllers/notee/application_controller.rb
@@ -3,23 +3,25 @@ module Notee
     before_action :restrict_access_json
 
     def restrict_access_json
-      redirect_to new_token_path && return unless confirm_expired_token
-      redirect_to new_token_path && return unless confirm_exist_token
+      return redirect_to new_token_path unless confirm_exist_token
+      return redirect_to new_token_path unless confirm_expired_token
     end
 
     private
 
     def confirm_exist_token
-      if Token.exists?(access_token: session[:access_token])
-        return true
-      else
+      unless Token.exists?(access_token: session[:access_token])
         session.delete(:access_token)
         return false
       end
+
+      true
     end
 
     def confirm_expired_token
       token = Token.find_by(access_token: session[:access_token])
+      return false unless token
+
       if Time.now > token.expires_at
         token.destroy
         session.delete(:access_token)
